refine acquire token logic
